prep site to move to PHP 7 sudo update-alternatives php

+++ modules/site/default/configurations_mc.php
@@ -5,19 +5,22 @@
 	
     $siteConfiguration = new \Site\Configuration();
     if (!empty($_REQUEST['key'])) $siteConfiguration = new \Site\Configuration($_REQUEST['key']);
-    switch ($_REQUEST['todo']) {
-        case 'drop':
-            $siteConfiguration->delete();
-            break;
+    if (isset($_REQUEST['todo'])) {
+        switch ($_REQUEST['todo']) {
+            case 'drop':
+                $siteConfiguration->delete();
+                break;
 
-        case 'add':
-            $siteConfiguration->key = $_REQUEST['key'];
-            $siteConfiguration->set($_REQUEST['value']);
-            break;
-            
-        case 'update':
-            $siteConfiguration->set($_REQUEST['value']);
-            break;
+            case 'add':
+                $siteConfiguration->key = $_REQUEST['key'];
+                $siteConfiguration->set($_REQUEST['value']);
+                break;
+                
+            case 'update':
+                $siteConfiguration->set($_REQUEST['value']);
+                break;
+        }
     }
+    
 	$siteConfigurations = new \Site\ConfigurationList();
 	$configuration = $siteConfigurations->find();
